src/helm-mcp-server/pyproject.toml
[project]
name = "helm-mcp-server"
version = "0.2.0"
description = "A Model Context Protocol (MCP) server for managing Kubernetes workloads via Helm."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.3",
    "kubernetes>=34.1.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.scripts]
helm-mcp-server = "helm_mcp_server.main:cli"

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helm_mcp_server"]

[tool.hatch.metadata]
allow-direct-references = true
